+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 #### 功能构建
 
+- 优化 `familyhelper-ui-view` 模块打包机制。
+  - 将打包机制由 `build -> lint` 更改为 `build -> check`，允许子模块引入更多的检查机制。
+  - 修改对应的构建脚本。
+
 - 建立 `familyhelper-ui-view` 模块构建脚本格式化机制。
 
 #### Bug修复